Added matches tries to save load system
Added current score, remaining tries and score to SaveLoad system.
Added remaining tries to game data

+++ Assets/Scripts/ScriptableObjects/CardScriptableObject.cs
@@ -7,6 +7,6 @@ public class CardScriptableObject : ScriptableObject
 {
     public GameObject foodItem;
     public FoodTypeEnum foodType;
-    public float amount = 10;
+    public int amount = 10;
     public Color bgColor;
 }
